Implement view core
To be used with my repo for ghc. You might not want to use it now as it
only contains the core needed to implement views and miss various
basic things. It may ruin your DB if used without care. Please wait some
days before it will work as expected by plan.